+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Release 5.5.0:
    - In `OpenId4VciClient.loadCredentialWithOffer()` remove parameter `requestedAttributes`
    - In `WalletService`, deprecate `CredentialRequestInput`
    - In `WalletService`, deprecate `createCredentialRequest(CredentialRequestInput)`, provide new method `createCredentialRequest(TokenResponseParameters)` for direct processing of the token response
+   - Iron out details for filtering scope and authorization details in `SimpleAuthorizationService`
    - Remove proof type `cwt`, which has been removed from draft 14
 
 Release 5.4.3: